add final fixes, small refactoring of decimalsToFixed in the places it's used with native tokens
